flex-algo: promote affinity-map + SRLG to global tables
Hard cutover of the named affinity (admin-group) and SRLG tables from
IS-IS-local config paths to top-level globals so OSPFv2/v3 can share
them. Config is broadcast to every protocol task, so each IGP keeps its
own copy fed from the same stream — no RIB-owned table or cross-task
plumbing is needed.

- Move AffinityMap and the SRLG types into crate::flex_algo
  (affinity_map.rs, srlg.rs) with config paths hardcoded to the new
  globals /affinity-map and /srlg. The AffinityBits impl moves with
  AffinityMap.
- isis::affinity_map / isis::srlg slim to thin shims: the IS-IS config
  callbacks (re-pathed to /affinity-map) + re-exports. inst.rs now
  dispatches /srlg/group instead of /router/isis/srlg/group.
- YANG: move the affinity-map list and srlg container out of
  `container isis` to top-level siblings of `container router`.

Breaking: existing /router/isis/affinity-map and /router/isis/srlg
configs must move to the global paths. No in-tree fixtures used the old
paths. Per-interface affinity/srlg leaf-lists stay under the IS-IS
interface subtree and resolve against the task-local copy unchanged.

yang_load guard + full zebra-rs unit suite (768) pass; workspace clippy
+ fmt clean.
