+++ zebra-rs/src/flex_algo/mod.rs
@@ -8,8 +8,12 @@
 //! sub-TLVs) and the config-callback shims stay in each protocol
 //! module; only the pure data model and constraint logic live here.
 
+pub mod affinity_map;
 pub mod constraint;
 pub mod entry;
+pub mod srlg;
 
+pub use affinity_map::AffinityMap;
 pub use constraint::{AffinityBits, link_passes_fad, local_link_affinity};
 pub use entry::{FadMetricType, FlexAlgoEntry};
+pub use srlg::{SrlgGroup, SrlgGroupBuilder};
